Fix some environment problems. Mess with test cases.

+++ src/tests.py
@@ -20,32 +20,36 @@ from makenovel import *
 
 class TestNovel(unittest.TestCase):
     
+    CURRDIR = os.path.abspath( os.path.dirname(__file__) )
+    
     novel = None
     env = None
-    CURRDIR = os.path.abspath( os.path.dirname(__file__) )
     proj_path = None
     
-    def setUp(self):
+    def __init__(self, methodName):
         
         self.proj_path = os.path.join(TestNovel.CURRDIR, "testnovel")
         
-        if os.path.exists(self.proj_path):
-            shutil.rmtree(self.proj_path)
+        config = Novel.getUserConfig()
             
         create_project("testnovel", "Test Novel", "ancient")
         
         title = "Test Novel"
-        config = {
-            "chapter.ext": Config("File extension for chapter files.", "rst"),
-            }
         author = Author("John", "Smith")
         
         env = NovelEnvironment(projdir = self.proj_path)
         
         self.novel = Novel(title, author, config, env)
-        self.novel.parts.extend([Part(self.novel), Part(self.novel),])
-        self.novel.plotlines.extend([Plotline(self.novel, "main", "Main Plotline"),
-                                 Plotline(self.novel, "side", "Side plot"),])
+        self.novel.parts = [Part(self.novel), Part(self.novel),]
+        self.novel.plotlines = [Plotline(self.novel, "main", "Main Plotline"),
+                                 Plotline(self.novel, "side", "Side plot"),]
+        self.novel.write_parts()
+        self.novel.write_plotlines()
+        self.novel.git_commit_data(Novel.PLOTLINESFILE,
+                                   message="Add test plotlines")
+        self.novel.git_commit_data(Novel.PARTSFILE, message='add test parts')
+        
+        super(TestNovel, self).__init__(methodName)
     
     def testBasicAdds(self):
         self.assertEqual(len(self.novel.parts), 2)
@@ -67,8 +71,8 @@ class TestNovel(unittest.TestCase):
     
     def testAddPlotline(self):
         self.assertEqual(len(self.novel.plotlines), 2)
-        print("PLF!!!!")
-        with open(PLOTLINESFILE) as plf:
+        pth = os.path.join(self.proj_path, Novel.PLOTLINESFILE)
+        with open(pth) as plf:
             for l in plf:
                 print("|%s" % l)
             plf.close()
@@ -83,11 +87,12 @@ class TestNovel(unittest.TestCase):
             self.assertEqual(self.novel.parts[i].number, i+1)
     
     def tearDown(self):
-        shutil.rmtree(self.proj_path)
+        if self.proj_path and os.path.exists(self.proj_path):
+            print("Removing %s" % self.proj_path)
+            shutil.rmtree(self.proj_path)
         self.novel = None
         self.env = None
         self.proj_path = None
-
 
 if __name__=='__main__':
     unittest.main()
